# docs/openapi/clipmato-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clipmato Public API",
    "summary": "Versioned public API contracts for uploads, progress, records, scheduling, and publishing.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/runtime/status": {
      "get": {
        "tags": [
          "Public API"
        ],
        "summary": "Runtime Status",
        "description": "Return the resolved runtime configuration and blockers.",
        "operationId": "runtime_status_api_v1_runtime_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeStatusModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/project-presets": {
      "get": {
        "tags": [
          "Public API"
        ],
        "summary": "List Project Presets",
        "description": "Return the saved project presets available to API clients.",
        "operationId": "list_project_presets_api_v1_project_presets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPresetListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/records": {
      "get": {
        "tags": [
          "Public API"
        ],
        "summary": "List Records",
        "description": "List records using the versioned public summary contract.",
        "operationId": "list_records_api_v1_records_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/record/{record_id}": {
      "get": {
        "tags": [
          "Public API"
        ],
        "summary": "Get Record",
        "description": "Return one record using the versioned public detail contract.",
        "operationId": "get_record_api_v1_record__record_id__get",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordDetailModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/progress/{record_id}": {
      "get": {
        "tags": [
          "Public API"
        ],
        "summary": "Get Progress",
        "description": "Return the current progress state for a record.",
        "operationId": "get_progress_api_v1_progress__record_id__get",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/record/{record_id}/title": {
      "post": {
        "tags": [
          "Public API"
        ],
        "summary": "Select Title",
        "description": "Persist the selected title for a record.",
        "operationId": "select_title_api_v1_record__record_id__title_post",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TitleSelectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TitleUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MachineErrorEnvelope": {
        "properties": {
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "error": {
            "$ref": "#/components/schemas/MachineError"
          }
        },
        "type": "object",
        "required": [
          "correlation_id",
          "error"
        ],
        "title": "MachineErrorEnvelope",
        "description": "Envelope for error responses that also echoes the correlation ID."
      },
      "ProgressStatusResponse": {
        "properties": {
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "stage",
          "progress"
        ],
        "title": "ProgressStatusResponse"
      },
...
      "ProjectPresetListResponse": {
        "properties": {
          "project_presets": {
            "items": {
              "$ref": "#/components/schemas/ProjectPresetModel"
            },
            "type": "array",
            "title": "Project Presets"
          }
        },
        "type": "object",
        "title": "ProjectPresetListResponse"
      },
...
      "RecordListResponse": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/RecordSummaryModel"
            },
            "type": "array",
            "title": "Records"
          }
        },
        "type": "object",
        "title": "RecordListResponse"
      },
...
      "TitleSelectionRequest": {
        "properties": {
          "selected_title": {
            "type": "string",
            "title": "Selected Title"
          }
        },
        "type": "object",
        "required": [
          "selected_title"
        ],
        "title": "TitleSelectionRequest"
      },
      "TitleUpdateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "selected_title": {
            "type": "string",
            "title": "Selected Title"
          }
        },
        "type": "object",
        "required": [
          "id",
          "selected_title"
        ],
        "title": "TitleUpdateResponse"
      },
...
    }
  }
}